updated rcube.cpp

+++ rubycube/Model/RCube.h
@@ -108,11 +108,11 @@ public:
 
     virtual RubiksCube &b2() = 0;
 
-    string getCornerColorString(uint8_t ind) const;
+    string getCornerColorString(int ind) const;
 
-    uint8_t getCornerIndex(uint8_t ind) const;
+    int getCornerIndex(int ind) const;
 
-    uint8_t getCornerOrientation(uint8_t ind) const;
+    int getCornerOrientation(int ind) const;
 
     
 };
